+++ rarbg_local/providers/__init__.py
@@ -1,3 +1,4 @@
+import asyncio
 import logging
 from collections.abc import Callable, Iterable
 from concurrent.futures import ThreadPoolExecutor
@@ -5,8 +6,10 @@ from queue import Empty, Queue
 from threading import Semaphore, current_thread
 from typing import TypeVar
 
+from ..models import ITorrent
 from ..types import ImdbId, TmdbId
-from .abc import Provider
+from ..utils import create_monitored_task
+from .abc import MovieProvider, Provider
 
 T = TypeVar('T')
 ProviderType = Callable[..., Iterable[T]]
@@ -73,17 +76,24 @@ async def search_for_tv(
             logger.exception('Unable to load [TV] from %s', provider)
 
 
-async def search_for_movie(imdb_id: ImdbId, tmdb_id: TmdbId):
-    from .abc import MovieProvider
+async def search_for_movie(
+    imdb_id: ImdbId, tmdb_id: TmdbId
+) -> tuple[list[asyncio.Future[None]], asyncio.Queue[ITorrent]]:
+    async def worker(provider: MovieProvider):
+        try:
+            async for result in provider.search_for_movie(imdb_id, tmdb_id):
+                output_queue.put_nowait(result)
+        except Exception:
+            logger.exception('Unable to load [MOVIE] from %s', provider)
 
+    tasks = []
+    output_queue = asyncio.Queue[ITorrent]()
     for provider in get_providers():
         if not isinstance(provider, MovieProvider):
             continue
-        try:
-            async for result in provider.search_for_movie(imdb_id, tmdb_id):
-                yield result
-        except Exception:
-            logger.exception('Unable to load [MOVIE] from %s', provider)
+
+        tasks.append(create_monitored_task(worker(provider), output_queue.put_nowait))
+    return tasks, output_queue
 
 
 async def main():
